requirements.txt
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
psycopg2-binary==2.9.9
SALib==1.4.7
scipy==1.11.4
seaborn==0.13.1
shapely==2.0.2
statsmodels==0.14.0
redis==5.0.1
docker==7.0.0
PyLaTeX==1.4.2
pdflatex==0.1.3
azure-communication-email==1.0.0